docs: break out Sphinx patches from config file.
