+++ docs/patches/__init__.py
@@ -1,0 +1,5 @@
+"""This directory contains routines to be patched into Sphinx
+
+These may be bugfixes or extensions of functionality. They are patched
+in by conf.py.
+"""
